requirements.txt
# breezometer-api-demo requirements
pandas
numpy
Flask
tweepy
pylint
autopep8
python-dotenv
